pyproject.toml
[project]
name = "task-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi-jwt-auth>=0.5.0",
    "fastapi[all,standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.4.0",
    "sqlalchemy[asyncio]>=2.0.40",
]
